nav General changed, new components created for the profile pic and AUXIE logo
