--- Client/src/API/data.js
@@ -1,173 +1,0 @@
-const data = [
-    {
-        id: 1,
-        name: 'Trabajo 1',
-        trabajo: 'Descripción del Trabajo 1',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 2,
-        name: 'Trabajo 2',
-        trabajo: 'Descripción del Trabajo 2',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 3,
-        name: 'Trabajo 3',
-        trabajo: 'Descripción del Trabajo 3',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 4,
-        name: 'Trabajo 4',
-        trabajo: 'Descripción del Trabajo 4',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 5,
-        name: 'Trabajo 5',
-        trabajo: 'Descripción del Trabajo 5',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 6,
-        name: 'Trabajo 6',
-        trabajo: 'Descripción del Trabajo 6',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 7,
-        name: 'Trabajo 7',
-        trabajo: 'Descripción del Trabajo 7',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 8,
-        name: 'Trabajo 8',
-        trabajo: 'Descripción del Trabajo 8',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 9,
-        name: 'Trabajo 9',
-        trabajo: 'Descripción del Trabajo 9',
-        status: 'cancelled',
-        price: 1500,
-    },
-    {
-        id: 10,
-        name: 'Trabajo 10',
-        trabajo: 'Descripción del Trabajo 10',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 11,
-        name: 'Trabajo 11',
-        trabajo: 'Descripción del Trabajo 11',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 12,
-        name: 'Trabajo 12',
-        trabajo: 'Descripción del Trabajo 12',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 13,
-        name: 'Trabajo 13',
-        trabajo: 'Descripción del Trabajo 13',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 14,
-        name: 'Trabajo 14',
-        trabajo: 'Descripción del Trabajo 14',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 15,
-        name: 'Trabajo 15',
-        trabajo: 'Descripción del Trabajo 15',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 16,
-        name: 'Trabajo 16',
-        trabajo: 'Descripción del Trabajo 16',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 17,
-        name: 'Trabajo 17',
-        trabajo: 'Descripción del Trabajo 17',
-        status: 'done',
-        price: 1500,
-    },
-    {
-        id: 18,
-        name: 'Trabajo 18',
-        trabajo: 'Descripción del Trabajo 18',
-        status: 'pending',
-        price: 1500,
-    },
-
-    {
-        id: 19,
-        name: 'Trabajo 19',
-        trabajo: 'Descripción del Trabajo 19',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 20,
-        name: 'Trabajo 20',
-        trabajo: 'Descripción del Trabajo 20',
-        status: 'cancelled',
-        price: 1500,
-    },
-    {
-        id: 21,
-        name: 'Trabajo 21',
-        trabajo: 'Descripción del Trabajo 21',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 22,
-        name: 'Trabajo 22',
-        trabajo: 'Descripción del Trabajo 22',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 23,
-        name: 'Trabajo 23',
-        trabajo: 'Descripción del Trabajo 23',
-        status: 'pending',
-        price: 1500,
-    },
-    {
-        id: 24,
-        name: 'Trabajo 24',
-        trabajo: 'Descripción del Trabajo 24',
-        status: 'done',
-        price: 1500,
-    },
-]
-
-export default data
