+++ Client/src/components/logo/logoClient/LogoCLient.jsx
@@ -1,0 +1,16 @@
+import style from './logoClient.module.scss'
+import { Link } from 'react-router-dom'
+
+import LogoAuxie2 from '../../../assets/Logos/logoAuxie2.svg'
+
+const LogoClient = () => {
+    return (
+        <div className={style.logo}>
+            <Link to={'/homeconsumer'}>
+                <img src={LogoAuxie2} alt="Logo Auxie" className={style.img} />
+            </Link>
+        </div>
+    )
+}
+
+export default LogoClient
